fluentsearch searchTag variable now dictates the pods returned
