+++ Wolfram.Fluent.Plugin/WolframOperation.cs
@@ -1,0 +1,11 @@
+using Blast.Core.Results;
+
+namespace Wolfram.Fluent.Plugin
+{
+    public class WolframOperation : SearchOperationBase
+    {
+        public WolframOperation() : base("Open in Wolfram Alpha", "", "")
+        {
+        }
+    }
+}
