+++ Wolfram.Fluent.Plugin/WolframTag.cs
@@ -1,0 +1,8 @@
+namespace Wolfram.Fluent.Plugin
+{
+    public enum WolframTag
+    {
+        Define,
+        Word
+    }
+}
